Delete errorhandler package, correct server error messages and fix Chrome date error.
