--- src/main/java/com/example/WebApplication/exceptionsAndErrors/ModifiedExceptionHandler.java
@@ -1,54 +1,0 @@
-package com.example.WebApplication.exceptionsAndErrors;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.ObjectError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import java.util.ArrayList;
-import java.util.List;
-
-@ControllerAdvice
-public class ModifiedExceptionHandler extends ResponseEntityExceptionHandler
-{
-    @ExceptionHandler(Exception.class)
-    public final ResponseEntity handleAllExceptions(Exception ex,
-                                                    WebRequest request)
-    {
-        List<String> details = new ArrayList<>();
-        details.add(ex.getLocalizedMessage());
-        ErrorResponse error = new ErrorResponse("Server Error", details);
-        return new ResponseEntity(error, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-
-    @ExceptionHandler(MemberNotFoundException.class)
-    public final ResponseEntity<Object> handleUserNotFoundException(MemberNotFoundException ex,
-                                                                    WebRequest request)
-    {
-        List<String> details = new ArrayList<>();
-        details.add(ex.getLocalizedMessage());
-        ErrorResponse error = new ErrorResponse("Record Not Found", details);
-        return new ResponseEntity(error, HttpStatus.NOT_FOUND);
-    }
-
-    @Override
-    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex,
-                                                                  HttpHeaders headers,
-                                                                  HttpStatus status,
-                                                                  WebRequest request)
-    {
-        List<String> fields = new ArrayList<>();
-        List<String> details = new ArrayList<>();
-        for(ObjectError error : ex.getBindingResult().getAllErrors()) {
-            String[] messageParts = error.getDefaultMessage().split("\\|");
-            fields.add(messageParts[0]);
-            details.add(messageParts[1]);
-        }
-        ErrorResponse error = new ErrorResponse("Validation Failed", fields, details);
-        return new ResponseEntity(error, HttpStatus.BAD_REQUEST);
-    }
-}
