+++ src/main/resources/templates/form.html
@@ -19,26 +19,31 @@
     <div>
         <form th:action="@{/api/v1/member}" th:object="${member}" id="membershipForm" method="POST">
             <div class="form-box">
+
                 <label for="firstName">Prénom</label>
                 <input type="text" th:field="*{firstName}" class="" name="firstName" id="firstName" placeholder="Prénom">
                 <div class="error-hint" th:if="${#fields.hasErrors('firstName')}" th:errors="*{firstName}">Name Error></div>
                 <div class="error-hint hidden">Ce champs est obligatoire.</div>
+
                 <label for="lastName">Nom</label>
                 <input type="text" th:field="*{lastName}" class="" name="lastName" id="lastName" placeholder="Nom">
-                <div class="error-hint hidden">Ce champs est obligatoire.</div>
                 <div class="error-hint" th:if="${#fields.hasErrors('lastName')}" th:errors="*{lastName}">Name Error></div>
+                <div class="error-hint hidden">Ce champs est obligatoire.</div>
+
                 <label for="birthDate">Date de naissance</label>
                 <input type="date" th:field="*{birthDate}" class="" name="birthDate" id="birthDate" placeholder="Date de naissance">
-                <div class="error-hint hidden">Ce champs est obligatoire.</div>
                 <div class="error-hint" th:if="${#fields.hasErrors('birthDate')}" th:errors="*{birthDate}">Name Error></div>
+                <div class="error-hint hidden">Ce champs est obligatoire.</div>
+
                 <label for="email">Adresse courriel</label>
                 <input type="email" th:field="*{email}" class="" name="email" id="email" placeholder="abc@abc.com">
-                <div class="error-hint hidden">Ce champs est obligatoire.</div>
                 <div class="error-hint" th:if="${#fields.hasErrors('email')}" th:errors="*{email}">Name Error></div>
+                <div class="error-hint hidden">Ce champs est obligatoire.</div>
+
                 <div class="radio-box">
                     <p name="gender-selection" id="gender">Genre:</p>
-                    <div class="error-hint hidden">Choisissez une option.</div>
                     <div class="error-hint" th:if="${#fields.hasErrors('gender')}" th:errors="*{gender}">Name Error></div>
+                    <div class="error-hint hidden">Choisissez une option.</div>
                     <input type="radio" name="gender" id="female" value="FEMALE">
                     <label for="female">Female</label><br>
                     <input type="radio" name="gender" id="male" value="MALE">
@@ -46,15 +51,17 @@
                     <input type="radio" name="gender" id="other" value="OTHER">
                     <label for="other">Other</label><br>
                 </div>
+
                 <div class="check-box">
                     <input type="checkbox" name="passedExam" id="passedExam" value="true">
                     <label for="passedExam">J'ai réussi l'examen de l'ordre.</label>
                 </div>
+
             </div>
         <hr />
             <button type="submit" name="submitForm" id="submitForm">Enregistrez</button>
             <button type="reset" name="resetForm" id="resetForm">Réinitialisez</button>
         </form>
     </div>
-    <script type="text/javascript" th:src="@{/js/script.js}"></script>
+    <script type="module" th:src="@{/js/script.js}"></script>
 </body>
